requirements.txt
# Core dependencies
requests>=2.28.0
python-whois>=0.7.3
phonenumbers>=8.12.0
beautifulsoup4>=4.9.3

# Audio processing
librosa>=0.9.2
sounddevice>=0.4.5
SpeechRecognition>=3.8.1
numpy>=1.21.0
scipy>=1.7.0

# Geolocation
folium>=0.12.1
geopy>=2.2.0
ip2geotools>=0.1.6

# AI/ML
transformers>=4.21.0
torch>=1.9.0
scikit-learn>=0.24.2

# GUI
PyQt5>=5.15.6
tkinter  # Usually comes with Python

# Utilities
tqdm>=4.64.0
python-dateutil>=2.8.2
